Call logging.basicConfig only within main
Merely importing a module shouldn't have the side-effect of changing
the global logging config, which can only be done once per process.
It's fragile and could lead to confusing behavior where log format
differs depending on the order of imports.

Since this is only intended to enable output for the CLI, let's
move it within the CLI's main method.