Get docs working via Travis & github pages
Basic skeleton of docs are in place.

Currently, the classes/methods appearing in "API reference" should
be considered as examples only.  I'm not sure exactly what should
go in there as it still seems a bit unclear what is and isn't
public API.  The main goal here is to get the doc building toolchain
up and running.