Remove public API for now
The API is too unstable at the moment and likely needs some
significant revision before we could commit to supporting that.
Let's rebrand this project as only a CLI tool for now, until we
have time to revise the API (probably meaning until we're ready
to actually *use* the API from outside of the tool).

Fixes #9