Get rid of 'profiles' in module_defaults association
Previously, 'profiles' was the key part to get the right modulemd_defaults
units to associate and unassociate. However, as the keyword in profiles
always include speical character '.'(dot), it leads to mongodb query failure.

Since while assmbling the lists of md_defaults_assoc and md_defaults_unassoc,
the 'profiles' has already been considered, we don't have to re-query with
it, but, because association and unassociation happens at the same time in
threads, it could delete the new unit if association happens before unassociation,
so, we serialize the modulemd_defaults units unassociation and association.