Include all module packages if no profile is defined
Packages for modules are included based on the package name present in
the repositories module profiles. This leads to missing packages in the
output repositories, as not every package required is included in module
profiles.

In this change, if no module profile is defined in the UBI
configuration, include every package present in the module artifacts.

Fix #81