Fix creation of source RPMs output set (#164)
Previously the method _get_rpm_output_set() stripped away
all modular packages and kept only the latest version of packages.

This seems to be too strict for source RPMs in some edge cases, at the
time when source RPMs output set is created, beacause the output set is
based on final output sets of binary and debug rpms that are already properly
filtered - they include only the latest versions of RPMs and non-modular
packages. Because of that we don't need to check again if the source RPM is
modular or not, or keep the latest version of SRPM. Only blacklisting
should happen. There are some edge cases when the original implementation
stripped some source RPMs that should be part of the output set. It's
probable that the edge cases happen due to data inconsistency especially
for modular packages.

We always need to include source RPMs of all binary and debug RPMs
unless they're blacklisted.