Make publication of repos non-blocking [RHELDST-6394] (#172)
Previously ubipop waited for each triplet of repos
(binary, debug, source) that were currently
in processing until publication of those repos
is finished. Only after publication finished,
ubipop was allowed to start processing another triplet.
As repo triplet is independent from other ones, it's
not required to do this in blocking way. So now
publication of repos is triggered at the same time
as before but waiting for all to finish is moved to
end of ubipop processing.