Fix TTL extraction [RHELDST-20510]
Previously the TTL extraction was implemented in such way that
it could potentially extract the wrong value, if the URL path contained
a component that also matched the regex.
This commit changes the re.match() to re.search(), which ensures it
will find the first occurrence of the match.
It also changes whitelist_externals to allowlist_externals in tox.ini