Make publishing of repos optional [RHELDST-21951]
After the rhel-ubi repos sync, this tool also took care of publishing
the affected ubi repos to the CDN. However, this did not work for Exodus.
It was decided that all publishes should be done via Pub, which is
already configured and works for Exodus.

This commit prepares this tool for removal of publishing by making it
possible to switch the publishing on/off as needed.

It also removes the intended deduplication of srpms in
`_get_pulp_actions_src_pkgs()` which was a bit confusing and also
unnecessary, since deduplication is done during manifest creation.
Finally, it removes some forgotten unused variables and renames some
variables so it's more understandable.